Persistent Priority
Keeps the selected priority persistent on the upload paperwork page
(saved in a 1 day cookie) so that users do not need to always change the
priority in a set of documents.

General code clean up, remove some console.log's that were not required
anymore.

new functions in base.js for handling cookies.